% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x, }{a data frame of items with different values and weights}

\item{W, }{the capacity of knapsack}
}
\value{
a list with best value and selected elements
}
\description{
Given data frame \code{x}, max weight \code{W}, give the best solution with brutal force of O(2^n)
}
\examples{
set.seed(42)
n<- 1000000
knapsack_objects <-data.frame(
 w=sample(1:4000, size = n, replace = TRUE),
 v=runif(n = n, 0, 10000))
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
