\name{FancyPackRLiU6-package}
\alias{FancyPackRLiU6-package}
\alias{FancyPackRLiU6}
\docType{package}
\title{
\packageTitle{FancyPackRLiU6}
}
\description{
\packageDescription{FancyPackRLiU6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FancyPackRLiU6}
\packageIndices{FancyPackRLiU6}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FancyPackRLiU6}

Maintainer: \packageMaintainer{FancyPackRLiU6}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
