% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x, }{a data frame of items with different values and weights}

\item{W, }{the capacity of knapsack}
}
\value{
a list with approximate solution with value and selected elements
}
\description{
Given data frame \code{x}, max weight \code{W}, give the approximate solution
}
\examples{
set.seed(42)
n<- 1000000
knapsack_objects <-data.frame(
 w=sample(1:4000, size = n, replace = TRUE),
 v=runif(n = n, 0, 10000))
greedy_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
