% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x, }{a data frame of items with different values and weights}

\item{W, }{the capacity of knapsack}
}
\value{
a list with best value and selected elements
}
\description{
Given data frame \code{x}, max weight \code{W}, give the best solution with dynamic programing of O(n*W)
}
\examples{
set.seed(42)
n<- 1000000
knapsack_objects <-data.frame(
 w=sample(1:4000, size = n, replace = TRUE),
 v=runif(n = n, 0, 10000))
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)

}
